final save
Ended up doing 3 pages okay-ish. Not my best work, but super educational. Material is fun to use. Somewhat tricky, but coding with it engages me fully.
Also, on these 3 pages there are of course minor errors here and here. It really isn't perfect. It's an exercise to see if Material can be used and well, yeah, kind of. Will get better with practice.